+++ src/app/main/todos/components/todo-list/todo-list.component.ts
@@ -2,6 +2,8 @@ import { Component, OnInit } from '@angular/core';
 import { selectTodoList } from '@todos/store/todolist/selector';
 import { Store } from '@ngrx/store';
 import { Todo } from '@shared/business-domain/model/Todo';
+import { MatDialog } from '@angular/material/dialog';
+import { AddTodoModalComponent } from '../add-todo-modal/add-todo-modal.component';
 
 @Component({
   selector: 'app-todo-list',
@@ -11,9 +13,13 @@ import { Todo } from '@shared/business-domain/model/Todo';
 export class TodoListComponent implements OnInit {
   public todos: Todo[];
 
-  constructor(private store: Store) {}
+  constructor(private store: Store, private dialog: MatDialog) {}
 
   ngOnInit(): void {
     this.store.select(selectTodoList).subscribe((todos) => (this.todos = todos));
+  }
+
+  onClickAdd(): void {
+    this.dialog.open(AddTodoModalComponent);
   }
 }