+++ src/app/main/todos/services/todo-http.service.spec.ts
@@ -1,8 +1,8 @@
 import { TestBed } from '@angular/core/testing';
 import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
 
-import { TodoHttpService } from '@todos/services/todo-http.service';
-import { Todo } from '@shared/business-domain/model/Todo';
+import { TodoHttpService, mockBackEnd } from '@todos/services/todo-http.service';
+import { DescribedTodo, Todo } from '@shared/business-domain/model/Todo';
 import { TodoState } from '@shared/business-domain/model/TodoState';
 
 describe('TodoHttpService', () => {
@@ -19,6 +19,7 @@ describe('TodoHttpService', () => {
   });
   afterEach(() => {
     httpTestingController.verify();
+    mockBackEnd.clear();
   });
 
   it('should be created', () => {
@@ -46,13 +47,21 @@ describe('TodoHttpService', () => {
   });
 
   describe('getTodo', () => {
-    it('should call back-end to update todo (it doesnt really...)', () => {
-      todoHttp.getTodos().subscribe(() => {});
+    it('should call back-end to get one todo (it doesnt really...)', () => {
+      mockBackEnd.set(1, new Todo('Laundry', TodoState.DONE, 1));
 
-      const testRequest = httpTestingController.expectOne('assets/mocks/todos.json');
-      testRequest.flush([new Todo('Laundry', TodoState.DONE, 1)]);
       todoHttp.getTodo(1).subscribe((todo) => {
         expect(todo.title).toBe('Laundry');
+      });
+    });
+  });
+
+  describe('createTodo', () => {
+    it('should call back-end to create todo (it doesnt really...)', () => {
+      const todoToCreate = new DescribedTodo('Laundry', TodoState.TODO, undefined, 'Need to wash clothes');
+
+      todoHttp.createTodo(todoToCreate).subscribe((createdTodo) => {
+        expect(createdTodo).toEqual({ ...todoToCreate, id: 1 });
       });
     });
   });