+++ src/app/main/todos/store/todolist/actions.ts
@@ -1,5 +1,6 @@
 import { createAction, props } from '@ngrx/store';
 import { Todo } from '@shared/business-domain/model/Todo';
+import { DescribedTodo } from '@shared/business-domain/model/Todo/DescribedTodo';
 
 export const startLoadingList = createAction('[TodoList] Start loading');
 export const listDataLoaded = createAction('[TodoList] Data loaded', props<{ todos: Todo[] }>());
@@ -15,3 +16,7 @@ export const startUpdateTodoState = createAction(
 );
 export const todoStateUpdated = createAction('[TodoList] Todo state updated', props<{ updatedTodo: Todo }>());
 export const todoStateUpdateFailed = createAction('[TodoList] Todo state update failed', props<{ error: string }>());
+
+export const createTodo = createAction('[TodoList] Start creating todo', props<{ todoToCreate: DescribedTodo }>());
+export const todoCreated = createAction('[TodoList] Todo created', props<{ createdTodo: DescribedTodo }>());
+export const todoCreationFailed = createAction('[TodoList] Todo failed to create', props<{ error: string }>());